Revert "Implementing lull reporting at bundle level processing (#29882)" (#30648)
This reverts commit ffe2dba532028cdbbb5bca9c374f0a2d756ee8bf.
